requirements.txt
ffmpeg-python==0.2.0
matplotlib==3.5.2
mediapipe==0.8.10.1
numpy==1.22.4
open3d==0.15.2
opencv-python==4.6.0.66
pyrender==0.1.45
torch==1.12.0
tqdm==4.64.0
trimesh==3.14.0